+++ relay-protocol/src/lib.rs
@@ -29,7 +29,6 @@ mod value;
 pub use self::annotated::*;
 pub use self::condition::RuleCondition;
 pub use self::impls::*;
-pub use self::macros::*;
 pub use self::meta::*;
 pub use self::size::*;
 pub use self::traits::*;