+++ relay-server/src/actors/processor.rs
@@ -24,22 +24,22 @@ use relay_dynamic_config::{
 };
 use relay_event_normalization::replay::{self, ReplayError};
 use relay_event_normalization::{
-    nel, ClockDriftProcessor, DynamicMeasurementsConfig, LightNormalizationConfig,
-    MeasurementsConfig, TransactionNameConfig,
+    nel, ClockDriftProcessor, DynamicMeasurementsConfig, MeasurementsConfig,
+    NormalizeProcessorConfig, TransactionNameConfig,
 };
 use relay_event_normalization::{GeoIpLookup, RawUserAgentInfo};
-use relay_event_schema::processor::{self, ProcessingAction, ProcessingState};
+use relay_event_schema::processor::{self, process_value, ProcessingAction, ProcessingState};
 use relay_event_schema::protocol::{
     Breadcrumb, ClientReport, Contexts, Csp, Event, EventType, ExpectCt, ExpectStaple, Hpkp,
-    IpAddr, LenientString, Metrics, NetworkReportError, OtelContext, RelayInfo, Replay,
-    SecurityReportType, SessionAggregates, SessionAttributes, SessionStatus, SessionUpdate,
+    IpAddr, LenientString, Metrics, NetworkReportError, OtelContext, ProfileContext, RelayInfo,
+    Replay, SecurityReportType, SessionAggregates, SessionAttributes, SessionStatus, SessionUpdate,
     Timestamp, TraceContext, UserReport, Values,
 };
 use relay_filter::FilterStatKey;
 use relay_metrics::aggregator::AggregatorConfig;
 use relay_metrics::{Bucket, MergeBuckets, MetricNamespace};
 use relay_pii::{scrub_graphql, PiiAttachmentsProcessor, PiiConfigError, PiiProcessor};
-use relay_profiling::ProfileError;
+use relay_profiling::{ProfileError, ProfileId};
 use relay_protocol::{Annotated, Array, Empty, FromValue, Object, Value};
 use relay_quotas::{DataCategory, ReasonCode};
 use relay_replays::recording::RecordingScrubber;
@@ -58,7 +58,7 @@ use {
     crate::actors::project_cache::UpdateRateLimits,
     crate::utils::{EnvelopeLimiter, MetricsLimiter},
     relay_event_normalization::{span, StoreConfig, StoreProcessor},
-    relay_event_schema::protocol::{ProfileContext, Span},
+    relay_event_schema::protocol::Span,
     relay_metrics::Aggregator,
     relay_quotas::{RateLimitingError, RedisRateLimiter},
     relay_redis::RedisPool,
@@ -314,8 +314,8 @@ struct ProcessEnvelopeState<'a> {
     /// The managed envelope before processing.
     managed_envelope: ManagedEnvelope,
 
-    /// Whether there is a profiling item in the envelope.
-    has_profile: bool,
+    /// The ID of the profile in the envelope, if a valid profile exists.
+    profile_id: Option<ProfileId>,
 
     /// Reservoir evaluator that we use for dynamic sampling.
     reservoir: ReservoirEvaluator<'a>,
@@ -1096,15 +1096,15 @@ impl EnvelopeProcessorService {
             .items()
             .filter(|item| item.ty() == &ItemType::Transaction)
             .count();
-        let mut found_profile = false;
+        let mut profile_id = None;
         state.managed_envelope.retain_items(|item| match item.ty() {
             // Drop profile without a transaction in the same envelope.
             ItemType::Profile if transaction_count == 0 => ItemAction::DropSilently,
             // First profile found in the envelope, we'll keep it if metadata are valid.
-            ItemType::Profile if !found_profile => {
+            ItemType::Profile if profile_id.is_none() => {
                 match relay_profiling::parse_metadata(&item.payload(), state.project_id) {
-                    Ok(_) => {
-                        found_profile = true;
+                    Ok(id) => {
+                        profile_id = Some(id);
                         ItemAction::Keep
                     }
                     Err(err) => ItemAction::Drop(Outcome::Invalid(DiscardReason::Profiling(
@@ -1118,7 +1118,24 @@ impl EnvelopeProcessorService {
             ))),
             _ => ItemAction::Keep,
         });
-        state.has_profile = found_profile;
+        state.profile_id = profile_id;
+    }
+
+    /// Transfers the profile ID from the profile item to the transaction item.
+    ///
+    /// If profile processing happens at a later stage, we remove the context again.
+    fn transfer_profile_id(&self, state: &mut ProcessEnvelopeState) {
+        if let Some(event) = state.event.value_mut() {
+            if event.ty.value() == Some(&EventType::Transaction) {
+                let contexts = event.contexts.get_or_insert_with(Contexts::new);
+                // If we found a profile, add its ID to the profile context on the transaction.
+                if let Some(profile_id) = state.profile_id {
+                    contexts.add(ProfileContext {
+                        profile_id: Annotated::new(profile_id),
+                    });
+                }
+            }
+        }
     }
 
     /// Normalize monitor check-ins and remove invalid ones.
@@ -1182,17 +1199,13 @@ impl EnvelopeProcessorService {
             }
             _ => ItemAction::Keep,
         });
-        if let Some(event) = state.event.value_mut() {
-            if event.ty.value() == Some(&EventType::Transaction) {
-                let contexts = event.contexts.get_or_insert_with(Contexts::new);
-                // If we found a profile, add its ID to the profile context on the transaction.
-                if let Some(profile_id) = found_profile_id {
-                    contexts.add(ProfileContext {
-                        profile_id: Annotated::new(profile_id),
-                    });
-                // If not, we delete the profile context.
-                } else {
-                    contexts.remove::<ProfileContext>();
+        if found_profile_id.is_none() {
+            // Remove profile context from event.
+            if let Some(event) = state.event.value_mut() {
+                if event.ty.value() == Some(&EventType::Transaction) {
+                    if let Some(contexts) = event.contexts.value_mut() {
+                        contexts.remove::<ProfileContext>();
+                    }
                 }
             }
         }
@@ -1393,7 +1406,7 @@ impl EnvelopeProcessorService {
             sampling_project_state,
             project_id,
             managed_envelope,
-            has_profile: false,
+            profile_id: None,
             reservoir,
         })
     }
@@ -2197,7 +2210,7 @@ impl EnvelopeProcessorService {
                         generic_tags: config.map(|c| c.tags.as_slice()).unwrap_or_default(),
                         transaction_from_dsc,
                         sampling_result: &state.sampling_result,
-                        has_profile: state.has_profile,
+                        has_profile: state.profile_id.is_some(),
                     };
 
                     state.extracted_metrics.extend(extractor.extract(event)?);
@@ -2677,7 +2690,7 @@ impl EnvelopeProcessorService {
             .aggregator_config_for(MetricNamespace::Transactions);
 
         utils::log_transaction_name_metrics(&mut state.event, |event| {
-            let config = LightNormalizationConfig {
+            let config = NormalizeProcessorConfig {
                 client_ip: client_ipaddr.as_ref(),
                 user_agent: RawUserAgentInfo {
                     user_agent: request_meta.user_agent(),
@@ -2723,8 +2736,12 @@ impl EnvelopeProcessorService {
             };
 
             metric!(timer(RelayTimers::EventProcessingLightNormalization), {
-                relay_event_normalization::light_normalize_event(event, config)
-                    .map_err(|_| ProcessingError::InvalidTransaction)
+                process_value(
+                    event,
+                    &mut relay_event_normalization::NormalizeProcessor::new(config),
+                    ProcessingState::root(),
+                )
+                .map_err(|_| ProcessingError::InvalidTransaction)
             })
         })?;
 
@@ -2756,6 +2773,7 @@ impl EnvelopeProcessorService {
             });
 
             self.extract_event(state)?;
+            self.transfer_profile_id(state);
 
             if_processing!({
                 self.process_unreal(state)?;
@@ -2783,7 +2801,6 @@ impl EnvelopeProcessorService {
 
         if_processing!({
             self.enforce_quotas(state)?;
-            // We need the event parsed in order to set the profile context on it
             self.process_profiles(state);
             self.process_check_ins(state);
         });
@@ -3127,9 +3144,12 @@ mod tests {
     use std::str::FromStr;
 
     use chrono::{DateTime, TimeZone, Utc};
+    use insta::assert_debug_snapshot;
     use relay_base_schema::metrics::{DurationUnit, MetricUnit};
     use relay_common::glob2::LazyGlob;
-    use relay_event_normalization::{MeasurementsConfig, RedactionRule, TransactionNameRule};
+    use relay_event_normalization::{
+        MeasurementsConfig, NormalizeProcessor, RedactionRule, TransactionNameRule,
+    };
     use relay_event_schema::protocol::{EventId, TransactionSource};
     use relay_pii::DataScrubbingConfig;
     use relay_protocol::RuleCondition;
@@ -3370,7 +3390,7 @@ mod tests {
                     outcome_aggregator.clone(),
                     test_store.clone(),
                 ),
-                has_profile: false,
+                profile_id: None,
                 event_metrics_extracted: false,
                 reservoir: dummy_reservoir(),
             }
@@ -3571,6 +3591,30 @@ mod tests {
             global_config,
             #[cfg(feature = "processing")]
             aggregator,
+        };
+
+        EnvelopeProcessorService {
+            global_config: Arc::default(),
+            inner: Arc::new(inner),
+        }
+    }
+
+    /// Creates test processor that can be initialized in sync tests.
+    fn create_test_processor_sync(config: Config) -> EnvelopeProcessorService {
+        let inner = InnerProcessor {
+            config: Arc::new(config),
+            envelope_manager: Addr::dummy(),
+            project_cache: Addr::dummy(),
+            outcome_aggregator: Addr::dummy(),
+            upstream_relay: Addr::dummy(),
+            #[cfg(feature = "processing")]
+            rate_limiter: None,
+            #[cfg(feature = "processing")]
+            redis_pool: None,
+            geoip_lookup: None,
+            global_config: Addr::dummy(),
+            #[cfg(feature = "processing")]
+            aggregator: Addr::dummy(),
         };
 
         EnvelopeProcessorService {
@@ -4164,7 +4208,7 @@ mod tests {
 
         relay_statsd::with_capturing_test_client(|| {
             utils::log_transaction_name_metrics(&mut event, |event| {
-                let config = LightNormalizationConfig {
+                let config = NormalizeProcessorConfig {
                     transaction_name_config: TransactionNameConfig {
                         rules: &[TransactionNameRule {
                             pattern: LazyGlob::new("/foo/*/**".to_owned()),
@@ -4176,7 +4220,11 @@ mod tests {
                     },
                     ..Default::default()
                 };
-                relay_event_normalization::light_normalize_event(event, config)
+                process_value(
+                    event,
+                    &mut NormalizeProcessor::new(config),
+                    ProcessingState::root(),
+                )
             })
             .unwrap();
         })
@@ -4436,5 +4484,96 @@ mod tests {
         );
 
         assert_eq!(get_sampling_match(res).sample_rate(), 0.4);
+    }
+
+    #[test]
+    fn test_profile_id_transfered() {
+        // Setup
+        let processor = create_test_processor_sync(Default::default());
+        let event_id = EventId::new();
+        let dsn = "https://e12d836b15bb49d7bbf99e64295d995b:@sentry.io/42"
+            .parse()
+            .unwrap();
+        let request_meta = RequestMeta::new(dsn);
+        let mut envelope = Envelope::from_request(Some(event_id), request_meta);
+
+        // Add a valid transaction item.
+        envelope.add_item({
+            let mut item = Item::new(ItemType::Transaction);
+
+            item.set_payload(
+                ContentType::Json,
+                r#"
+            {
+                "type": "transaction",
+                "transaction": "/foo/",
+                "timestamp": 946684810.0,
+                "start_timestamp": 946684800.0,
+                "contexts": {
+                    "trace": {
+                    "trace_id": "4c79f60c11214eb38604f4ae0781bfb2",
+                    "span_id": "fa90fdead5f74053",
+                    "op": "http.server",
+                    "type": "trace"
+                    }
+                },
+                "transaction_info": {
+                    "source": "url"
+                }
+            }
+            "#,
+            );
+            item
+        });
+
+        // Add a profile to the same envelope.
+        envelope.add_item({
+            let mut item = Item::new(ItemType::Profile);
+            item.set_payload(
+                ContentType::Json,
+                r#"{
+                    "profile_id": "012d836b15bb49d7bbf99e64295d995b",
+                    "version": "1",
+                    "platform": "android",
+                    "os": {"name": "foo", "version": "bar"},
+                    "device": {"architecture": "zap"},
+                    "timestamp": "2023-10-10 00:00:00Z"
+                }"#,
+            );
+            item
+        });
+
+        let mut project_state = ProjectState::allowed();
+        project_state.config.features.0.insert(Feature::Profiling);
+
+        let message = ProcessEnvelope {
+            envelope: ManagedEnvelope::standalone(envelope, Addr::dummy(), Addr::dummy()),
+            project_state: Arc::new(project_state),
+            sampling_project_state: None,
+            reservoir_counters: ReservoirCounters::default(),
+        };
+
+        let envelope_response = processor.process(message).unwrap();
+        let ctx = envelope_response.envelope.unwrap();
+        let new_envelope = ctx.envelope();
+
+        // Get the re-serialized context.
+        let item = new_envelope
+            .get_item_by(|item| item.ty() == &ItemType::Transaction)
+            .unwrap();
+        let transaction = Annotated::<Event>::from_json_bytes(&item.payload()).unwrap();
+        let context = transaction
+            .value()
+            .unwrap()
+            .context::<ProfileContext>()
+            .unwrap();
+
+        assert_debug_snapshot!(context, @r###"
+        ProfileContext {
+            profile_id: EventId(
+                012d836b-15bb-49d7-bbf9-9e64295d995b,
+            ),
+        }
+        "###);
     }
 }